requirements.txt
pyTelegramBotAPI==4.16.1
python-dotenv==0.21.0
pymongo==4.6.3
requests==2.31.0